requirements.txt
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
decorator==4.4.2
docopt==0.6.2
ez-setup==0.9
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
peewee==3.17.3
pillow==10.3.0
proglog==0.1.10
pytube==15.0.0
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
tqdm==4.66.2
urllib3==2.2.1
webvtt-py==0.4.6
